Making the code more readable
I added some comments to the code that I worked on with Proshanto last week. I attempted to do more work on the updateModel() function to increase the amount of pacMan movement allowed, though once the y-coordinates of the pacMan go above 3 the function(s) seem to break. Updating Main soon.