Adding a ghost
I began to add functionality to a ghost turtle that should chase the player around the screen using only the four cardinal directions(?). There are big differences between this code and the code I made for my CS project. Also, I've run into an error where the coordinates of the ghost(x specifically) are not associated with a value when the pacMan turtle occupies the same coordinates. I'd like to ask Proshanto how to go about debugging this in an effective manner, as adding a condition to each if statement in movePacMan() does not seem ideal to me.