Updating Main and ReadMe
I was able to ask Proshanto about the issue I was having when the ghost and PacMan entered the same coordinate point in the 2d list. He was able to help me fix it. I updated Main as I felt that it was outdated. Additionally, I updated the ReadMe for potential future readers.